O2 f 1008 product create review consumer/update review aggregate

+++ src/main/java/com/dailyon/productservice/product/dto/response/ReadProductDetailResponse.java
@@ -23,7 +23,7 @@ public class ReadProductDetailResponse {
     private Integer price;
     private Gender gender;
 
-    private Float avgRating;
+    private Double avgRating;
     private Long reviewCount;
 
     private List<ReadProductStockResponse> productStocks;
